requirements.txt
atlassian-python-api>=3.41.4
notion-client>=2.2.1
python-dotenv>=1.0.0
requests>=2.31.0 